WIP: Rename PubspecReader to ConfigReader
